requirements.txt
pillow
tqdm
torch
torchvision
